Split test_ray_init while enabling cluster tests in it
